+++ core/src/main/java/com/graphhopper/routing/InstructionsFromEdges.java
@@ -73,6 +73,7 @@ public class InstructionsFromEdges implements Path.EdgeVisitor {
     private double prevInstructionPrevOrientation = Double.NaN;
     private Instruction prevInstruction;
     private boolean prevInRoundabout;
+    private String prevDestinationAndRef;
     private String prevName;
     private String prevInstructionName;
 
@@ -138,17 +139,23 @@ public class InstructionsFromEdges implements Path.EdgeVisitor {
         }
 
         final String name = (String) edge.getValue("name");
+        final String ref = (String) edge.getValue("ref");
+        final String destination = (String) edge.getValue("destination"); // getValue is fast if it does not exist in edge
+        final String destinationRef = (String) edge.getValue("destination_ref");
         if ((prevInstruction == null) && (!isRoundabout)) // very first instruction (if not in Roundabout)
         {
             int sign = Instruction.CONTINUE_ON_STREET;
             prevInstruction = new Instruction(sign, name, new PointList(10, nodeAccess.is3D()));
-            prevInstruction.setExtraInfo("ref", edge.getValue("ref"));
+            prevInstruction.setExtraInfo("ref", ref);
+            prevInstruction.setExtraInfo("destination", destination);
+            prevInstruction.setExtraInfo("destination_ref", destinationRef);
             double startLat = nodeAccess.getLat(baseNode);
             double startLon = nodeAccess.getLon(baseNode);
             double heading = AngleCalc.ANGLE_CALC.calcAzimuth(startLat, startLon, latitude, longitude);
             prevInstruction.setExtraInfo("heading", Helper.round(heading, 2));
             ways.add(prevInstruction);
             prevName = name;
+            prevDestinationAndRef = destination + destinationRef;
 
         } else if (isRoundabout) {
             // remark: names and annotations within roundabout are ignored
@@ -182,6 +189,7 @@ public class InstructionsFromEdges implements Path.EdgeVisitor {
                 {
                     prevOrientation = AngleCalc.ANGLE_CALC.calcOrientation(prevLat, prevLon, latitude, longitude);
                     prevName = name;
+                    prevDestinationAndRef = destination + destinationRef;
                 }
                 prevInstruction = roundaboutInstruction;
                 ways.add(prevInstruction);
@@ -200,7 +208,9 @@ public class InstructionsFromEdges implements Path.EdgeVisitor {
         } else if (prevInRoundabout) //previously in roundabout but not anymore
         {
             prevInstruction.setName(name);
-            prevInstruction.setExtraInfo("ref", edge.getValue("ref"));
+            prevInstruction.setExtraInfo("ref", ref);
+            prevInstruction.setExtraInfo("destination", destination);
+            prevInstruction.setExtraInfo("destination_ref", destinationRef);
 
             // calc angle between roundabout entrance and exit
             double orientation = AngleCalc.ANGLE_CALC.calcOrientation(prevLat, prevLon, latitude, longitude);
@@ -220,10 +230,10 @@ public class InstructionsFromEdges implements Path.EdgeVisitor {
 
             prevInstructionName = prevName;
             prevName = name;
+            prevDestinationAndRef = destination + destinationRef;
 
         } else {
-            int sign = getTurn(edge, baseNode, prevNode, adjNode, name);
-
+            int sign = getTurn(edge, baseNode, prevNode, adjNode, name, destination + destinationRef);
             if (sign != Instruction.IGNORE) {
                 /*
                     Check if the next instruction is likely to only be a short connector to execute a u-turn
@@ -274,11 +284,14 @@ public class InstructionsFromEdges implements Path.EdgeVisitor {
                     prevInstructionName = prevName;
                     ways.add(prevInstruction);
                 }
-                prevInstruction.setExtraInfo("ref", edge.getValue("ref"));
+                prevInstruction.setExtraInfo("ref", ref);
+                prevInstruction.setExtraInfo("destination", destination);
+                prevInstruction.setExtraInfo("destination_ref", destinationRef);
             }
             // Update the prevName, since we don't always create an instruction on name changes the previous
             // name can be an old name. This leads to incorrect turn instructions due to name changes
             prevName = name;
+            prevDestinationAndRef = destination + destinationRef;
         }
 
         updatePointsAndInstruction(edge, wayGeo);
@@ -316,7 +329,7 @@ public class InstructionsFromEdges implements Path.EdgeVisitor {
         ways.add(finishInstruction);
     }
 
-    private int getTurn(EdgeIteratorState edge, int baseNode, int prevNode, int adjNode, String name) {
+    private int getTurn(EdgeIteratorState edge, int baseNode, int prevNode, int adjNode, String name, String destinationAndRef) {
         GHPoint point = InstructionsHelper.getPointForOrientationCalculation(edge, nodeAccess);
         double lat = point.getLat();
         double lon = point.getLon();
@@ -377,6 +390,7 @@ public class InstructionsFromEdges implements Path.EdgeVisitor {
         if (otherContinue != null) {
             // We are at a fork
             if (!InstructionsHelper.isNameSimilar(name, prevName)
+                    || !InstructionsHelper.isNameSimilar(destinationAndRef, prevDestinationAndRef)
                     || InstructionsHelper.isNameSimilar(otherContinue.getName(), prevName)
                     || !outgoingEdgesAreSlower) {
 
@@ -410,13 +424,9 @@ public class InstructionsFromEdges implements Path.EdgeVisitor {
             }
         }
 
-        if (!outgoingEdgesAreSlower) {
-            if (Math.abs(delta) > .6
-                    || outgoingEdges.isLeavingCurrentStreet(prevName, name)) {
-                // Leave the current road -> create instruction
-                return sign;
-
-            }
+        if (!outgoingEdgesAreSlower && (Math.abs(delta) > .6 || outgoingEdges.isLeavingCurrentStreet(prevName, name))) {
+            // Leave the current road -> create instruction
+            return sign;
         }
 
         return Instruction.IGNORE;
@@ -430,8 +440,10 @@ public class InstructionsFromEdges implements Path.EdgeVisitor {
         }
         double newDist = edge.getDistance();
         prevInstruction.setDistance(newDist + prevInstruction.getDistance());
-        // todo: why do we not account for turn times here ?
-        prevInstruction.setTime(weighting.calcEdgeMillis(edge, false) + prevInstruction.getTime());
+        if (prevEdge != null)
+            prevInstruction.setTime(GHUtility.calcMillisWithTurnMillis(weighting, edge, false, prevEdge.getEdge()) + prevInstruction.getTime());
+        else
+            prevInstruction.setTime(weighting.calcEdgeMillis(edge, false) + prevInstruction.getTime());
     }
 
 }